Refactor getPanes into its own Panes component
